Add -debugbacktrace flag for printing debugging backtraces
The flag allows us to easily debug program execution. This is only enabled
n Linux currently (using GLIBC), but should be doable for other operating
systems as well.

Added a preliminary Backtrace::output() call in CBlock::CheckBlock.

It seems the only makefile under src/ that has actually been maintained up
to this point is the Linux one - so no changes have been made in the other
ones for now.