Parallelize WalletModel::checkBalanceChanged()
This fixes the first serious issues causing hangups in the UI when we have
a big amount of transactions active in the wallet. The problem is caused
by the wallet balance functions looping through the transaction list
several times on each call to checkBalanceChanged().

Implemented by issuing a boost::signal that creates a asynchronous thread
via std::async.