Completely rework how fixed nodes are added and handled
The previous solution used a python script that referenced a text file
containing nodes. This script then generated a chainparamsseeds.h file
using that information.

This whole thing seems messy, exaggerated and complicated. This reworks
the solution by just statically adding the fixed nodes directly in
chainparams.cpp, in the same way as the DNS seeds are added.