Certain MacOS X flags are now only set when Xcode is found
This when the wallet is built on a native system. For cross compilation,
these flags need to be set by the build system via a generated qmake
configuration file.

While we are at it, we move the requirement from OSX 10.7 to OSX 10.11.
There is really no point in supporting any older version at this point.